+++ index.js
@@ -3,10 +3,11 @@
 import "./features/commands/mainCommand";
 
 // Dungeons
-import "./features/dungeons/dungeonParty";
+import "./features/dungeons/dungeonPartyInfo";
 
 // Crimson Isles
 import "./features/crimsonIsles/mobHitboxes";
+import "./features/crimsonIsles/kuudraPartyInfo"
 
 // Misc
 import "./features/misc/damageRenderer";
